+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## v0.9.0
 
-### Major changes 
+### Major changes
 
-The `draw()` function no longer needs `Self` to be mutable - by @PPakalns in [#18](https://github.com/Blatko1/wgpu-text/pull/18).
+The internal `draw()` function no longer needs `Self` to be mutable - by @PPakalns in [#18](https://github.com/Blatko1/wgpu-text/pull/18).
+
+The `BrushError` enum is now public.
 
 ## v0.8.3
