--- src/internal/mod.rs
@@ -1,3 +1,0 @@
-mod brush;
-mod cache;
-mod pipeline;
