+++ src/pipeline.rs
@@ -2,7 +2,7 @@ use glyph_brush::Rectangle;
 use std::num::NonZeroU32;
 use wgpu::util::DeviceExt;
 
-use crate::{internal::cache::Cache, vertex::Vertex, Matrix};
+use crate::{cache::Cache, Matrix};
 
 /// Responsible for drawing text.
 #[derive(Debug)]
@@ -28,7 +28,7 @@ impl Pipeline {
         let cache = Cache::new(device, tex_dimensions, matrix);
 
         let shader =
-            device.create_shader_module(wgpu::include_wgsl!("../shader/shader.wgsl"));
+            device.create_shader_module(wgpu::include_wgsl!("shader/shader.wgsl"));
 
         let vertex_buffer = device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("wgpu-text Vertex Buffer"),
@@ -129,5 +129,102 @@ impl Pipeline {
     #[inline]
     pub fn resize_texture(&mut self, device: &wgpu::Device, tex_dimensions: (u32, u32)) {
         self.cache.recreate_texture(device, tex_dimensions);
+    }
+}
+
+use glyph_brush::ab_glyph::{point, Rect};
+
+#[repr(C)]
+#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
+pub struct Vertex {
+    top_left: [f32; 3],
+    bottom_right: [f32; 2],
+    tex_top_left: [f32; 2],
+    tex_bottom_right: [f32; 2],
+    color: [f32; 4],
+}
+
+impl Vertex {
+    pub fn to_vertex(
+        glyph_brush::GlyphVertex {
+            mut tex_coords,
+            pixel_coords,
+            bounds,
+            extra,
+        }: glyph_brush::GlyphVertex,
+    ) -> Vertex {
+        let bounds = bounds;
+        let mut rect = Rect {
+            min: point(pixel_coords.min.x, pixel_coords.min.y),
+            max: point(pixel_coords.max.x, pixel_coords.max.y),
+        };
+
+        // handle overlapping bounds, modify uv_rect to preserve texture aspect
+        if rect.max.x > bounds.max.x {
+            let old_width = rect.width();
+            rect.max.x = bounds.max.x;
+            tex_coords.max.x =
+                tex_coords.min.x + tex_coords.width() * rect.width() / old_width;
+        }
+        if rect.min.x < bounds.min.x {
+            let old_width = rect.width();
+            rect.min.x = bounds.min.x;
+            tex_coords.min.x =
+                tex_coords.max.x - tex_coords.width() * rect.width() / old_width;
+        }
+        if rect.max.y > bounds.max.y {
+            let old_height = rect.height();
+            rect.max.y = bounds.max.y;
+            tex_coords.max.y =
+                tex_coords.min.y + tex_coords.height() * rect.height() / old_height;
+        }
+        if rect.min.y < bounds.min.y {
+            let old_height = rect.height();
+            rect.min.y = bounds.min.y;
+            tex_coords.min.y =
+                tex_coords.max.y - tex_coords.height() * rect.height() / old_height;
+        }
+
+        Vertex {
+            top_left: [rect.min.x, rect.min.y, extra.z],
+            bottom_right: [rect.max.x, rect.max.y],
+            tex_top_left: [tex_coords.min.x, tex_coords.min.y],
+            tex_bottom_right: [tex_coords.max.x, tex_coords.max.y],
+            color: extra.color,
+        }
+    }
+
+    pub fn buffer_layout() -> wgpu::VertexBufferLayout<'static> {
+        wgpu::VertexBufferLayout {
+            array_stride: std::mem::size_of::<Self>() as wgpu::BufferAddress,
+            step_mode: wgpu::VertexStepMode::Instance,
+            attributes: &[
+                wgpu::VertexAttribute {
+                    format: wgpu::VertexFormat::Float32x3,
+                    offset: 0,
+                    shader_location: 0,
+                },
+                wgpu::VertexAttribute {
+                    format: wgpu::VertexFormat::Float32x2,
+                    offset: std::mem::size_of::<[f32; 3]>() as wgpu::BufferAddress,
+                    shader_location: 1,
+                },
+                wgpu::VertexAttribute {
+                    format: wgpu::VertexFormat::Float32x2,
+                    offset: std::mem::size_of::<[f32; 5]>() as wgpu::BufferAddress,
+                    shader_location: 2,
+                },
+                wgpu::VertexAttribute {
+                    format: wgpu::VertexFormat::Float32x2,
+                    offset: std::mem::size_of::<[f32; 7]>() as wgpu::BufferAddress,
+                    shader_location: 3,
+                },
+                wgpu::VertexAttribute {
+                    format: wgpu::VertexFormat::Float32x4,
+                    offset: std::mem::size_of::<[f32; 9]>() as wgpu::BufferAddress,
+                    shader_location: 4,
+                },
+            ],
+        }
     }
 }
